Add files via upload
The dataset to be used for this project is for the apartments in Dubai. The dataset has 1905 observations from 38 variables. Dubai has grown to become one of the key financial markets in the Middle East, and there is so much interest in its real estate markets for both commercial purposes and residentials. Here, I am hoping to draw meaningful insights that could advise the key stakeholders of the factors that influence the pricing of properties in Dubai. For instance, properties on the Dubai’s exotic Palm Jumeirah could attract higher prices while properties that are distant from the malls and metro stations could be valued lower.
I will try to answer the question; what are the determining factors of rising prices of apartments in Dubai? 
Which area affects prices the most? What are the properties worth in Dubai?